+++ src/components/Header/Header.tsx
@@ -2,8 +2,7 @@ import Container from '@/ui/Container';
 import Logo from '@/ui/Logo';
 import Navbar from '../Navbar';
 import HeaderTitle from './HeaderTitle';
-import LoginButton from '@/features/auth/LoginButton';
-import SignUpButton from '@/features/auth/SignUpButton';
+import AuthControls from '@/features/auth/AuthControls';
 
 function Header() {
   return (
@@ -14,10 +13,7 @@ function Header() {
             <Logo />
             <HeaderTitle />
           </div>
-          <div className='space-x-2 md:space-x-4'>
-            <LoginButton />
-            <SignUpButton />
-          </div>
+          <AuthControls />
         </Container>
       </div>
       <div className='bg-base-300 px-4 py-2 lg:px-6 lg:py-4'>