+++ src/features/auth/LoginButton/LoginButton.tsx
@@ -1,9 +1,0 @@
-function LoginButton() {
-  return (
-    <button className='btn btn-sm btn-accent md:btn-md tracking-wider'>
-      Login
-    </button>
-  );
-}
-
-export default LoginButton;