+++ src/features/auth/SignUpButton/SignUpButton.tsx
@@ -1,9 +1,0 @@
-function SignUpButton() {
-  return (
-    <button className='btn btn-sm btn-accent btn-outline md:btn-md tracking-wider'>
-      Sign up
-    </button>
-  );
-}
-
-export default SignUpButton;